refactor: move code from to_cid into cid
As there is not custom `ToCid` trait anymore, move the `TryFrom`
implementations to the main `cid` module.

Except for moving code, this commit doesn't contain any other changes.
